+++ src/Mailer.php
@@ -16,8 +16,7 @@ namespace Berlioz\Mailer;
 
 use Berlioz\Mailer\Exception\InvalidArgumentException;
 use Berlioz\Mailer\Exception\TransportException;
-use Berlioz\Mailer\Transport\PhpMail as MailTransport;
-use Berlioz\Mailer\Transport\TransportInterface;
+use Berlioz\Mailer\Transport;
 use Psr\Log\LoggerAwareInterface;
 use Psr\Log\LoggerInterface;
 use ReflectionClass;
@@ -27,10 +26,10 @@ class Mailer implements LoggerAwareInterface
 {
     /** Default transports into the package. */
     public const DEFAULT_TRANSPORTS = [
-        'smtp' => '\Berlioz\Mailer\Transport\Smtp',
-        'mail' => '\Berlioz\Mailer\Transport\PhpMail'
+        'smtp' => Transport\Smtp::class,
+        'mail' => Transport\PhpMail::class,
     ];
-    /** @var TransportInterface Transport */
+    /** @var Transport\TransportInterface Transport */
     private $transport;
     /** @var LoggerInterface The logger instance. */
     private $logger;
@@ -88,9 +87,9 @@ class Mailer implements LoggerAwareInterface
             $class = new ReflectionClass($className);
             $object = $class->newInstanceArgs($classArgs);
 
-            if (!$object instanceof TransportInterface) {
+            if (!$object instanceof Transport\TransportInterface) {
                 throw new InvalidArgumentException(
-                    'Transport class must be an instance of \Berlioz\Mailer\Transport\TransportInterface interface'
+                    sprintf('Transport class must be an instance of %s interface', Transport\TransportInterface::class)
                 );
             }
 
@@ -117,12 +116,12 @@ class Mailer implements LoggerAwareInterface
     /**
      * Get transport.
      *
-     * @return TransportInterface
+     * @return Transport\TransportInterface
      */
-    public function getTransport(): TransportInterface
+    public function getTransport(): Transport\TransportInterface
     {
         if (null === $this->transport) {
-            $this->transport = new MailTransport();
+            $this->transport = new Transport\PhpMail();
         }
 
         return $this->transport;
@@ -131,11 +130,11 @@ class Mailer implements LoggerAwareInterface
     /**
      * Set transport.
      *
-     * @param TransportInterface $transport
+     * @param Transport\TransportInterface $transport
      *
      * @return static
      */
-    public function setTransport(TransportInterface $transport): Mailer
+    public function setTransport(Transport\TransportInterface $transport): Mailer
     {
         // Logger
         if ($transport instanceof LoggerAwareInterface && null !== $this->logger) {
@@ -165,7 +164,7 @@ class Mailer implements LoggerAwareInterface
      *
      * @param Mail $mail Mail
      * @param Address[] $addresses Address list
-     * @param callable $callback Callback called after each email sent
+     * @param callable|null $callback Callback called after each email sent
      *
      * @return mixed Depends of transport
      * @throws TransportException if an error occurred during sending of mail.